JUnit test AddLifeInsuranceCommandTest and utility class LifeInsuranceBuilder

+++ src/main/java/seedu/address/model/insurance/UniqueLifeInsuranceList.java
@@ -46,12 +46,13 @@ public class UniqueLifeInsuranceList implements Iterable<LifeInsurance> {
             if (insurance1.getPremium().equals(insurance2.getPremium())) {
                 return 0;
             } else {
-                return insurance1.getPremium() < insurance2.getPremium() ? 1 : -1;
+                return insurance1.getPremium().toDouble() < insurance2.getPremium().toDouble() ? 1 : -1;
             }
         });
     }
     //@@author
 
+    //@@author OscarWang114
     /**
      * Returns the backing list as an unmodifiable {@code ObservableList}.
      */
